requirements/base.txt
# Base requirements for all environments
Django==5.2.3
wagtail==7.0.1
psycopg[binary]==3.2.9
pillow==11.3.0
python-decouple==3.8
whitenoise==6.9.0
